+++ src/cardMarkup.ts
@@ -1,19 +1,12 @@
 import { MarkdownRenderChild, MarkdownRenderer } from "obsidian";
-import {
-	isDefined,
-	createElementShort,
-	dateTokenConfigurationIsTypeNumber,
-	dateTokenConfigurationIsTypeString,
-} from "~/utils";
+import { isDefined, createElementShort } from "~/utils";
 
 import type {
 	MarkdownCodeBlockTimelineProcessingContext,
 	CardContent,
 	AutoTimelineSettings,
-	AbstractDate,
-	DateTokenConfiguration,
-	DateTokenType,
 } from "~/types";
+import { formatAbstractDate } from "./abstractDateFormating";
 
 /**
  * Generates a card in the DOM based on given ccontext.
@@ -110,6 +103,7 @@ export function formatBodyForCard(body?: string | null): string {
 }
 
 /**
+ * Get the text displayed in the card where the date should be.
  *
  * @param param0 - The context for a single card.
  * @param param0.startDate - the start date of an event.
@@ -126,104 +120,4 @@ export function getDateText(
 	if (!isDefined(endDate)) return formatedStart;
 
 	return `From ${formatedStart} to ${formatAbstractDate(endDate, settings)}`;
-}
-
-/**
- *
- * @param date - Target date to format.
- * @param param1 - The settings of the plugin.
- * @param param1.dateDisplayFormat - The target format to displat the date in.
- * @param param1.dateParserGroupPriority - The token priority list for the date format.
- * @param param1.dateTokenConfiguration - The configuration for the given date format.
- * @returns the formated representation of a given date based off the plugins settings.
- */
-export function formatAbstractDate(
-	date: AbstractDate | boolean,
-	{
-		dateDisplayFormat,
-		dateParserGroupPriority,
-		dateTokenConfiguration,
-	}: Pick<
-		AutoTimelineSettings,
-		| "dateDisplayFormat"
-		| "dateParserGroupPriority"
-		| "dateTokenConfiguration"
-	>
-): string {
-	if (typeof date === "boolean") return "now";
-	const prioArray = dateParserGroupPriority.split(",");
-	let output = dateDisplayFormat.toString();
-
-	prioArray.forEach((token, index) => {
-		const configuration = dateTokenConfiguration.find(
-			({ name }) => name === token
-		);
-
-		if (!configuration)
-			throw new Error(
-				`[April's not so automatic timelines] - No date token configuration found for ${token}, please setup your date tokens correctly`
-			);
-
-		output = output.replace(
-			`{${token}}`,
-			formatDateToken(date[index], configuration)
-		);
-	});
-
-	return output;
-}
-
-/**
- * Shorthand to format a part of an abstract date.
- *
- * @param datePart - fragment of an abstract date.
- * @param configuration - the configuration bound to that date token.
- * @returns the formated token.
- */
-export function formatDateToken(
-	datePart: number,
-	configuration: DateTokenConfiguration
-): string {
-	if (dateTokenConfigurationIsTypeNumber(configuration))
-		return formatNumberDateToken(datePart, configuration);
-	if (dateTokenConfigurationIsTypeString(configuration))
-		return formatStringDateToken(datePart, configuration);
-	throw new Error(
-		`[April's not so automatic timelines] - Corrupted date token configuration, please reset settings`
-	);
-}
-
-/**
- * Used to quickly format a fragment of an abstract date based off a number typed date token configuration.
- *
- * @param datePart - fragment of an abstract date.
- * @param param1 - A numerical date token configuration to apply.
- * @param param1.minLeght - the minimal length of a numerical date input.
- * @returns the formated token.
- */
-function formatNumberDateToken(
-	datePart: number,
-	{ minLeght }: DateTokenConfiguration<DateTokenType.number>
-): string {
-	let stringifiedToken = datePart.toString();
-
-	if (minLeght < 0) return stringifiedToken;
-	while (stringifiedToken.length < minLeght)
-		stringifiedToken = "0" + stringifiedToken;
-	return stringifiedToken;
-}
-
-/**
- * Used to quickly format a fragment of an abstract date based off a string typed date token configuration.
- *
- * @param datePart - fragment of an abstract date.
- * @param param1 - A string typed date token configuration to apply.
- * @param param1.dictionary - the relation dictionary for a date string typed token.
- * @returns the formated token.
- */
-function formatStringDateToken(
-	datePart: number,
-	{ dictionary }: DateTokenConfiguration<DateTokenType.string>
-): string {
-	return dictionary[datePart];
 }
